requirements.txt
# Core requirements
streamlit>=1.32.0
langchain-chroma>=0.1.0
langchain-openai>=0.1.0
langchain-huggingface>=0.0.1
langchain-core>=0.1.0

# Azure & Data Ingestion
azure-storage-blob>=12.19.0
sec-edgar-downloader>=3.0.0
beautifulsoup4>=4.12.3
markdownify>=0.11.6

# Environment Management
python-dotenv>=1.0.1
